started explore.tsx page and utils/omni_tx fetchCollection
