+++ .eslintrc.js
@@ -6,7 +6,6 @@ module.exports = {
     'eslint:recommended',
     'plugin:@typescript-eslint/recommended',
     'plugin:@next/next/recommended',
-    'prettier',
   ],
   rules: {
     semi: ['error', 'never'],
