+++ components/Web3Status.tsx
@@ -12,6 +12,7 @@ import { getContract } from '../utils/contracts.js'
 import { useWeb3React } from '@web3-react/core'
 import { InjectedConnector } from '@web3-react/injected-connector'
 import { BigNumber, ethers } from 'ethers'
+import { Web3Interface } from '../utils/utils'
 
 import { ToastContainer, toast } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
@@ -83,7 +84,7 @@ const addresses = {
 
 
 export default function Web3Status() {
-  const {chainId, activate, account, active, connector, error } = useWeb3React()
+  const {chainId, activate, account, active, connector, error } = useWeb3React() as Web3Interface
 
 
   const [toChain, setToChain] = useState('4')
@@ -101,218 +102,6 @@ export default function Web3Status() {
       errorToast('connect error')
     }
   }
-
-  // async function disconnect() {
-  //   try {
-  //     deactivate()
-  //   } catch (ex) {
-  //   }
-  // }
-
-  // const checkConnect = () => {
-  //   let keys = Object.keys(addresses)
-  //   let flag = keys.filter(item => {
-  //     if(item == chainId?.toString()) {
-  //       return item
-  //     }
-  //   })
-
-  //   if(flag.length == 0) {
-  //     errorToast('Please select correct Network!')
-  //     return false
-  //   }
-  //   return true
-  // }
-
-  // useEffect(() => {
-  //   if(chainId!=undefined){
-  //     {
-  //         switchNetwork()
-  //     }
-  //   }
-  // }, [netId])
-
-  // useEffect(()=>{
-  //   if(isSwitching==true){
-  //     if(account!=undefined){
-  //       getInfo()
-  //       setIsSwitching(false)
-  //     }
-  //   }
-  // },[isSwitching,account])
-
-  // useEffect (()=>{
-  //   Aos.init({ duration: 1000 })
-  //   if(chainId==undefined){
-  //     switchNetwork()
-  //   } 
-  // }, [])
-
-  // const switchNetwork = async () => {
-  //   const provider = window.ethereum as any
-  //   if(chainId?.toString() != netId){
-  //     try {
-  //       await provider.request({
-  //         method: 'wallet_switchEthereumChain',
-  //         params: [
-  //           {
-  //             chainId: `0x${Number(netId).toString(16)}`,
-  //           }
-  //         ]      
-  //       })
-        
-  //       setTransferNFT()
-  //       setIsSwitching(true)
-  //     } catch (addError) {
-  //       if(addError["code"]==4001){
-  //         errorToast('User rejected the request to switch network')
-  //         setNetId(chainId)
-  //       } else{
-  //         if(netId=='4'){
-  //           await window.ethereum.request({
-  //             method: 'wallet_addEthereumChain',
-  //             params: [
-  //               {
-  //                 chainId: '0x4',
-  //                 chainName: 'Rinkeby Test Network',
-  //                 nativeCurrency: {
-  //                     name: 'rinkeby',
-  //                     symbol: 'ETH',
-  //                     decimals: 18
-  //               },
-  //               rpcUrls: ['https://rinkeby.infura.io/v3/'],
-  //               blockExplorerUrls: ['https://rinkeby.etherscan.io']
-  //               }
-  //             ]
-  //           })
-  //           setTransferNFT()
-  //           setIsSwitching(true)
-  //         } else if (netId=='97'){
-  //           await window.ethereum.request({
-  //             method: 'wallet_addEthereumChain',
-  //             params: [
-  //               {
-  //                 chainId: '0x61',
-  //                 chainName: 'BNB Smart Chain Testnet',
-  //                 nativeCurrency: {
-  //                     name: 'Binance Coin',
-  //                     symbol: 'tBNB',
-  //                     decimals: 18
-  //               },
-  //               rpcUrls: ['https://data-seed-prebsc-1-s1.binance.org:8545/'],
-  //               blockExplorerUrls: ['https://testnet.bscscan.com/']
-  //               }
-  //             ]
-  //           })
-  //           setTransferNFT()
-  //           setIsSwitching(true)
-  //         } else if(netId=='43113'){
-  //           await window.ethereum.request({
-  //             method: 'wallet_addEthereumChain',
-  //             params: [
-  //               {
-  //                 chainId: '0xA869',
-  //                 chainName: 'Avalanche Testnet C-Chain',
-  //                 nativeCurrency: {
-  //                     name: 'Avalanche',
-  //                     symbol: 'AVAX',
-  //                     decimals: 18
-  //               },
-  //               rpcUrls: ['https://api.avax-test.network/ext/bc/C/rpc'],
-  //               blockExplorerUrls: ['https://testnet.snowtrace.io/']
-  //               }
-  //             ]
-  //           })
-  //           setTransferNFT()
-  //           setIsSwitching(true)
-  //         } else if(netId=='80001'){
-  //           await window.ethereum.request({
-  //             method: 'wallet_addEthereumChain',
-  //             params: [
-  //               {
-  //                 chainId: '0x13881',
-  //                 chainName: 'Mumbai Testnet',
-  //                 nativeCurrency: {
-  //                     name: 'Mumbai',
-  //                     symbol: 'MATIC',
-  //                     decimals: 18
-  //               },
-  //               rpcUrls: ['https://rpc-mumbai.maticvigil.com/'],
-  //               blockExplorerUrls: ['https://polygonscan.com/']
-  //               }
-  //             ]
-  //           })
-  //           setTransferNFT()
-  //           setIsSwitching(true)
-  //         } else if(netId=='421611'){
-  //           await window.ethereum.request({
-  //             method: 'wallet_addEthereumChain',
-  //             params: [
-  //               {
-  //                 chainId: '0x66EEB',
-  //                 chainName: 'Arbitrum Rinkeby',
-  //                 nativeCurrency: {
-  //                     name: 'Arbitrum',
-  //                     symbol: 'ETH',
-  //                     decimals: 18
-  //               },
-  //               rpcUrls: ['https://rinkeby.arbitrum.io/rpc'],
-  //               blockExplorerUrls: ['https://testnet.arbiscan.io/']
-  //               }
-  //             ]
-  //           })
-  //           setTransferNFT()
-  //           setIsSwitching(true)
-  //         } 
-  //         else if(netId=='4002'){
-  //           await window.ethereum.request({
-  //             method: 'wallet_addEthereumChain',
-  //             params: [
-  //               {
-  //                 chainId: '0xFA2',
-  //                 chainName: 'Fantom testnet',
-  //                 nativeCurrency: {
-  //                     name: 'Fantom',
-  //                     symbol: 'MTF',
-  //                     decimals: 18
-  //               },
-  //               rpcUrls: ['https://rpc.testnet.fantom.network/'],
-  //               blockExplorerUrls: ['https://faucet.fantom.network/']
-  //               }
-  //             ]
-  //           })
-  //           setTransferNFT()
-  //           setIsSwitching(true)
-  //         } else if(netId=='69'){
-  //           await window.ethereum.request({
-  //             method: 'wallet_addEthereumChain',
-  //             params: [
-  //               {
-  //                 chainId: '0x45',
-  //                 chainName: 'Optimistic Ethereum Testnet Kovan',
-  //                 nativeCurrency: {
-  //                     name: 'Kovan',
-  //                     symbol: 'KOR',
-  //                     decimals: 18
-  //               },
-  //               rpcUrls: ['https://kovan.optimism.io/'],
-  //               blockExplorerUrls: ['']
-  //               }
-  //             ]
-  //           })
-  //           setTransferNFT()
-  //           setIsSwitching(true)
-  //         }else{
-  //           setNetId(chainId)
-  //           errorToast('Switching network error, please try again')
-  //         }
-  //       }
-        
-  //     }
-  //   } else {
-  //     setIsSwitching(true)
-  //   }
-  // }
 
 
   const loadingIcon = () => {
