+++ package.json
@@ -11,6 +11,9 @@
   "dependencies": {
     "@web3-react/core": "^6.1.9",
     "@web3-react/injected-connector": "^6.0.7",
+    "aos": "^2.3.4",
+    "autoprefixer": "^10.4.7",
+    "eip-712": "^1.0.0",
     "ethers": "^5.6.8",
     "next": "12.1.6",
     "react": "18.1.0",
